+++ seed/utils/goals.py
@@ -2,8 +2,6 @@
 SEED Platform (TM), Copyright (c) Alliance for Sustainable Energy, LLC, and other contributors.
 See also https://github.com/SEED-platform/seed/blob/main/LICENSE.md
 """
-import math
-from pint import Quantity
 
 from django.db.models import Case, F, FloatField, IntegerField, Prefetch, Sum, Value, When
 from django.db.models.fields.json import KeyTextTransform
@@ -216,37 +214,3 @@ def get_state_pairs(property_ids, goal_id):
         state_pairs.append({"property": property, "baseline": baseline_state, "current": current_state})
 
     return state_pairs
-
-def combine_properties(p1, p2):
-    if not p2:
-        return p1
-    combined = p1.copy()
-    for key, value in p2.items():
-        if value is not None:
-            combined[key] = value
-    return combined
-
-
-def percentage(a, b):
-    if not a or b is None:
-        return None
-    value = round(((a - b) / a) * 100)
-    return None if math.isnan(value) else value
-
-
-def get_preferred(p, columns):
-    if not p:
-        return
-    for col in columns:
-        return convert_quantity(p[col])
-
-
-def convert_quantity(value):
-    if isinstance(value, Quantity):
-        value = value.m
-    return value
-
-
-def get_kbtu(prop, key):
-    if prop[f"{key}_sqft"] is not None and prop[f"{key}_eui"] is not None:
-        return round(prop[f"{key}_sqft"] * prop[f"{key}_eui"])
