+++ seed/views/v3/goals.py
@@ -9,6 +9,7 @@ from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
 from django.db.utils import DataError
 from django.http import JsonResponse
 from django.utils.decorators import method_decorator
+from pint import Quantity
 from rest_framework import status
 from rest_framework.decorators import action
 
@@ -20,7 +21,7 @@ from seed.serializers.pint import apply_display_unit_preferences
 from seed.utils.api import OrgMixin
 from seed.utils.api_schema import swagger_auto_schema_org_query_param
 from seed.utils.goal_notes import get_permission_data
-from seed.utils.goals import get_or_create_goal_notes, get_portfolio_summary, combine_properties, percentage, get_preferred, convert_quantity, get_kbtu
+from seed.utils.goals import get_or_create_goal_notes, get_portfolio_summary
 from seed.utils.search import FilterError, build_view_filters_and_sorts, filter_views_on_related
 from seed.utils.viewsets import ModelViewSetWithoutPatch
 
@@ -293,3 +294,37 @@ class GoalViewSet(ModelViewSetWithoutPatch, OrgMixin):
             }
         )
 
+
+def combine_properties(p1, p2):
+    if not p2:
+        return p1
+    combined = p1.copy()
+    for key, value in p2.items():
+        if value is not None:
+            combined[key] = value
+    return combined
+
+
+def percentage(a, b):
+    if not a or b is None:
+        return None
+    value = round(((a - b) / a) * 100)
+    return None if math.isnan(value) else value
+
+
+def get_preferred(p, columns):
+    if not p:
+        return
+    for col in columns:
+        return convert_quantity(p[col])
+
+
+def convert_quantity(value):
+    if isinstance(value, Quantity):
+        value = value.m
+    return value
+
+
+def get_kbtu(prop, key):
+    if prop[f"{key}_sqft"] is not None and prop[f"{key}_eui"] is not None:
+        return round(prop[f"{key}_sqft"] * prop[f"{key}_eui"])
